+++ src/es/udc/cartolab/gvsig/epanet/structures/JunctionWrapper.java
@@ -12,14 +12,14 @@ public class JunctionWrapper extends NodeWrapper {
 	super(iFeature);
     }
 
-    public JunctionWrapper(String id, double x, double y, int elevation,
-	    int baseDemand) {
+    public JunctionWrapper(String id, double x, double y, double elevation,
+	    double baseDemand) {
 	super();
 	createJunction(id, x, y, elevation, baseDemand);
     }
 
-    protected void createJunction(String id, double x, double y, int elevation,
-	    int baseDemand) {
+    protected void createJunction(String id, double x, double y,
+	    double elevation, double baseDemand) {
 	// TODO: Probablemente habría que hacer dos métodos, 1 para los nodos
 	// con demanda y otro para los que no tiene demanda. Chequear
 	// parseJunction
